Short and Long Calculator
It is 1 upper level than basics. It is easy to code but it's very long and mind boggling code. It will confuse the beginners as to where the code gets wrong. I tried it 7 times to get the short calculator get done. Then long one is easy but very long coding required. It took me 1 hour to code both long and short code in one go except my cases of 7 times trying.. 
Try it once.