Short Calculator to Short Calculator
It is for beginners but they had to try again and again to get the result. Getting the a and b enter value tab is easy. But getting the add/sub/mul/div answer after writing is somehow difficult. There will be syntax error every time you try. Just writing the names add, sub, div, mul is difficult to get but if you are writing 1,2,3,4 instead of that than it's easy in the "loop" python code. 
But if you get once how to write than you won't forgot ever after..