Grade and Percentage Calculator
It's a beginner level code. 
It will tell the percentage for 12th class commerce with maths and arts only. 
You can edit for your subjects. 
I also edit some joke for fun as you are trying on your own to learn some code so I thought of adding fun things in between..