Create Square Root Finder
It is a very simple code. 
It will find square root of a number. 
But basically some people having problem in it like how to put 1/2 on the top . 
So it's simple just use 0.5 instead of it. 
And remember when using '**' , your code will calculate from right to left. 
This is the only case in which right to left calculation is done..