Create Fahrenheit to Celcius Calculator and vice versa
It is just a Fahrenheit to Celcius and vice versa calculator. 
But basically there is something new. 
There is one thing tough in it. 
That how can we do Fahrenheit to Celcius and Celcius to Fahrenheit without knowing that the user is putting which value. 
So just ask the user anyhow you like. 
I asked by putting option of 1 and 2
It's your choice you can even use 'F' and 'C'. 
And the rest of calculation code is mostly same as area of circle code.